// src/app/model/CauHoi/CauHoi_LuaChon_DTO.model.ts
export class CauHoi_LuaChon_DTO{
    maCauHoi: number;
    noiDung: string;
    goiYcauHoi: string;
    maLoaiCauHoi: number;
    maNhomCauHoi: number;
    nguoiThem: string;
    ngayThem: Date;
    ngaySua: Date;
    nguoiSua: string;
    trangThai:number;
    maLuaChon: number;
    maCauHoi_LC: number;
    noiDung_LC: string;
    trangThai_LC: number;

}